requirements.txt
langchain
langchain_community
langchain_huggingface
langchain_groq
faiss-cpu
langchain_core
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers
